+++ examples/pasts.rs
@@ -1,17 +1,18 @@
-use core::{cell::Cell, pin::pin};
+use core::pin::pin;
 
 use pasts::{notify, prelude::*, Executor};
-use shared_cell::CellExt;
+use shared_cell::shared_cell;
 
 struct Context {
     stuff: u32,
 }
 
 fn main() {
-    // A `Cell` is enough to share data between async tasks on the same thread
-    let context = Cell::new(Context { stuff: 12 });
-
     Executor::default().block_on(async move {
+        let context = Context { stuff: 12 };
+
+        shared_cell!(context);
+
         let task_a = pin!(async { context.with(|context| context.stuff += 2) });
         let task_b = pin!(async { context.with(|context| context.stuff -= 1) });
         let (task_a, task_b) = (&mut task_a.fuse(), &mut task_b.fuse());